# apps/server/config/swagger.yml
swagger: '2.0'
info:
  description: >-
    Quizzer is a web application that can be used for Quiz nights with your
    friends, colleagues or even strangers. A Quizz night consits of two parties:
    a Quizz master and the teams that want to compete. The Quizz master is the
    all-powerfull leader of the Quizz night. He uses a tablet determine what
    categories will be played, what questions will be answered, how many rounds
    will be played and even score the answers. Teams can choose their own name
    and must work together to achieve victory. A team plays together on one
    SmartPhone. Together they submit their answer to the current question of the
    round. After each round points are awarded to top three teams who have the
    most correct answers in the round. Scores are displayed real time on a big
    screen. There teams can see the current score, number of correct answers and
    the ranking of each team.
  version: 0.0.1
  title: Quizzer
  termsOfService: 'https://github.com/HANICA-DWA/fall2017-quizz-SharonEnRick'
  contact:
    email: rick_en_sharon_quizzer@rickvanlieshout.com
  license:
    name: MIT
    url: 'https://choosealicense.com/licenses/mit/'
host: localhost
basePath: /
schemes:
  - http
...
paths:
...
  '/quizzes/{id}/answers/{teamName}/':
    put:
      tags:
        - Answers
      summary: Add / update an answer
      description: Update or add a new answer
      operationId: addAnswer
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id of the quiz
          required: true
          type: string
        - name: teamName
          in: path
          description: teamname submitting the answer
          required: true
          type: string
        - in: body
          name: body
          description: ''
          required: true
          schema:
            $ref: '#/definitions/GivenAnswer'
      responses:
        '200':
          description: answer submitted
...
  '/quizzes/{id}/judge/':
    put:
      summary: judge the current question
      description: judge the current question
      operationId: judgeQuestion
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id of the quiz
          required: true
          type: string
        - in: body
          name: body
          description: ''
          required: true
          schema:
            $ref: '#/definitions/Judgement'
      responses:
        '200':
          description: answer submitted
  '/quizzes/{id}/startQuiz':
    post:
      tags:
        - Quizzes
      summary: Starts the quiz
      description: Starts the quiz with approved teams
      operationId: startQuiz
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id of the quiz
          required: true
          type: string
        - in: body
          name: body
          description: quizId and teams object for starting the quiz
          required: true
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: quiz started
...
definitions:
...
  GivenAnswer:
    description: Answers given
    type: object
    properties:
      answer:
        type: string
...
  Judgement:
    description: judgement
    type: object
    properties:
      teamId:
        type: string
      judgement:
        type: boolean
...